mslot_class: find_free_tfi(): use uint32_t to shift 1 << 31
Avoid runtime error seen on jenkins admin-deb9build slave, when building
osmo-pcu with_dsp=None,with_vty=False, during ts_alloc and tbf regression
tests:

+../../../src/mslot_class.c:242:22: runtime error: left shift of 1 by 31 places cannot be represented in type 'int'

The master-osmo-pcu breaks on this since moving to a debian9 build slave.

+++ src/mslot_class.c
@@ -239,7 +239,7 @@ int8_t find_free_tfi(uint32_t tfi_map)
 		return -1;
 
 	for (tfi = 0; tfi < 32; tfi++) {
-		if (!(tfi_map & (1 << tfi)))
+		if (!(tfi_map & (((uint32_t)1) << tfi)))
 			return tfi;
 	}
